refactor: extract shared reddit_scraper package; thin bot/api entrypoints
Move logic out of the two monolithic entrypoints into a layered package so
bot.py and api.py stop duplicating config/credential/path logic:

  reddit_scraper/
    config.py        # data paths (env-aware), search.json access, source order
    credentials.py   # load/sanitize/encoding-guard/authenticate, CREDENTIALS state
    status.py        # bot_status.json writer
    notifications.py # apprise dispatch + notify_error
    sources.py       # oauth/rss/json fetchers + dispatcher, throttle, cooldown
    health.py        # Uptime Kuma heartbeats
    monitor.py       # RedditMonitor

- bot.py 1024 -> 127 lines (wiring + main loop), re-exports the names tests use
- api.py load_config/save_config/load_credentials/save_credentials/clean_monitor
  now delegate to the package (single source of truth); routes unchanged
- paths resolve from DATA_DIR at call time so reload-based tests stay correct
- bot.py no longer does blocking/auth work at import -> un-skipped test_bot.py
- .gitignore: ignore __pycache__/.DS_Store/.pytest_cache

Tests: 29 passed (was 26 passed / 3 skipped).
